udpate

+++ pipeline/calcs.py
@@ -109,16 +109,17 @@ def get_stage(stage, stage_num, ath_ids, gsheet="No", orange_df=prologue):
                 push_gsheet(f_df_o_pts, stage_num)
                 stage_res = pull_gsheet(stage_num)
 
+                stage_res['KOM'] = stage_res['KOM'].astype('int64')
+                stage_res['Int. S'] = stage_res['Int. S'].astype('int64')
+                stage_res['DS/DC'] = stage_res['DS/DC'].astype('int64')
+                stage_res['Report'] = stage_res['Report'].astype('int64')
+                stage_res['MAR'] = stage_res['MAR'].astype('int64')
+
                 path = live_race_path + stage_num + '.csv'
                 save_csv(stage_res, path )
             else:
                 stage_res = load_csv(path)
-                
-            stage_res['KOM'] = stage_res['KOM'].astype('int64')
-            stage_res['Int. S'] = stage_res['Int. S'].astype('int64')
-            stage_res['DS/DC'] = stage_res['DS/DC'].astype('int64')
-            stage_res['Report'] = stage_res['Report'].astype('int64')
-            stage_res['MAR'] = stage_res['MAR'].astype('int64')
+            
 
 
     return stage_res, orange_df
